Move logic for deleting old output and copying in image files into a new helper class
