+++ Source/Program.cs
@@ -19,7 +19,7 @@ namespace AssetGenerator
 
             var executingAssembly = Assembly.GetExecutingAssembly();
             var executingAssemblyFolder = Path.GetDirectoryName(executingAssembly.Location);
-            var imageFolder = Path.Combine(executingAssemblyFolder, "ImageDependencies");
+            //var imageFolder = Path.Combine(executingAssemblyFolder, "ImageDependencies");
 
             // Uses Reflection to create a list containing one instance of each test class 
             List<dynamic> testBatch = new List<dynamic>();
@@ -37,53 +37,13 @@ namespace AssetGenerator
             foreach (var test in testBatch)
             {
                 TestValues makeTest = new TestValues();
-                var combos = ComboHelper.AttributeCombos(test);
+                List<List<Property>> combos = ComboHelper.AttributeCombos(test);
                 LogBuilder logs = new LogBuilder();
+                string assetFolder = Path.Combine(executingAssemblyFolder, test.testType.ToString());
 
-                // Delete any preexisting files in the output directories, then create those directories if needed
-                var assetFolder = Path.Combine(executingAssemblyFolder, test.testType.ToString());
-                var trashFolder = Path.Combine(executingAssemblyFolder, "Delete");
-                bool tryAgain = true;
-                while (tryAgain)
-                {
-                    try
-                    {
-                        Directory.Move(assetFolder, trashFolder);
-                        Directory.Delete(trashFolder, true);
-                        tryAgain = false;
-                    }
-                    catch (DirectoryNotFoundException)
-                    {
-                        // Do nothing
-                        tryAgain = false;
-                    }
-                    catch (IOException)
-                    {
-                        Console.WriteLine("Unable to delete the directory.");
-                        Console.WriteLine("Verify that there are no open files and that the current user has write permission to that directory.");
-                        Console.WriteLine("Press any key to try again.");
-                        Console.ReadKey();
-                        tryAgain = true;
-                    }
-                }
-
+                FileHelper.ClearOldFiles(executingAssemblyFolder, assetFolder);
                 Directory.CreateDirectory(assetFolder);
-                ImageAttribute[] usedImages = (ImageAttribute[])Attribute.GetCustomAttributes(test.GetType(), typeof(ImageAttribute));
-                if (usedImages != null)
-                {
-                    foreach (var image in usedImages)
-                    {
-                        if (File.Exists(Path.Combine(imageFolder, image.Name)))
-                        {
-                            File.Copy(Path.Combine(imageFolder, image.Name), Path.Combine(assetFolder, image.Name), true);
-                        }
-                        else
-                        {
-                            Debug.WriteLine(imageFolder + " does not exist");
-                        }
-                    }
-                }
-
+                FileHelper.CopyImageFiles(executingAssemblyFolder, assetFolder, test.GetType());
                 logs.SetupHeader(test);
 
                 int numCombos = combos.Count;
